content/svg/content/test/test_animLengthRelativeUnits.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=508206
-->
<head>
  <title>Test for liveness of relative units in animation</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=508206">Mozilla Bug 508206</a>
<p id="display"></p>
<!-- XXX The following should be display: none but that's broken by bug 413975
     where we don't handle percentage lengths when the whole fragment is
     display: none properly. -->
<div id="content" style="visibility: hidden">
<svg id="svg" xmlns="http://www.w3.org/2000/svg" width="100px" height="100px"
     onload="this.pauseAnimations()">
  <g font-size="10px">
    <circle cx="0" cy="0" r="15" fill="blue" id="circle">
      <animate attributeName="cx" from="0" to="10em" dur="10s" begin="0s"
        fill="freeze" id="animate"/>
      <animate attributeName="cy" from="0" to="100%" dur="10s" begin="0s"
        fill="freeze"/>
    </circle>
  </g>
</svg>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[
/** Test liveness of relative units of animated lengths **/

/* Global Variables */
const svgns="http://www.w3.org/2000/svg";
var svg = document.getElementById("svg");
var circle = document.getElementById('circle');

SimpleTest.waitForExplicitFinish();

function main() {
  ok(svg.animationsPaused(), "should be paused by <svg> load handler");
  is(svg.getCurrentTime(), 0, "should be paused at 0 in <svg> load handler");

  // Sample mid-way through the animation
  svg.setCurrentTime(5);

  // (1) Check values mid-way
  is(circle.cx.animVal.value, 50,
    "(1) Unexpected animVal for cx before changing base length");
  is(circle.cy.animVal.value, 50,
    "(1) Unexpected animVal for cy before changing base length");

  // (2) Change the frame of reference and check values are updated immediately

  // Change font-size
  circle.parentNode.setAttribute('font-size', '5px');
  todo_is(circle.cx.animVal.value, 25,
    "(2) Unexpected animVal for cx after changing parent font-size");

  // Change the viewport size
  svg.setAttribute('height', '50px');
  todo_is(circle.cy.animVal.value, 25,
    "(2) Unexpected animVal for cy after changing containing viewport size");

  SimpleTest.finish();
}

var animate = document.getElementById('animate');
if (animate && animate.targetElement) {
  window.addEventListener("load", main, false);
} else {
  ok(true); // Skip tests but don't report 'todo' either
  SimpleTest.finish();
}
]]>
</script>
</pre>
</body>
</html>
